fix: remove hydration errors and refactor home page
