--- src/data/footer-quick-links.ts
@@ -1,16 +1,0 @@
-export const footerQuickLinks = [
-  {
-    text: "Canvas",
-    href: "https://chalmers.instructure.com/",
-  },
-  {
-    text: "TimeEdit",
-    href: "https://cloud.timeedit.net/chalmers/web/public/",
-    blank: true,
-  },
-  {
-    text: "Boka Grupprum",
-    href: "https://cloud.timeedit.net/chalmers/web/b1/",
-    blank: true,
-  },
-];
