+++ src/data/home-page-carousel-image-urls.ts
@@ -1,0 +1,5 @@
+export const homePageCarouselImageUrls: string[] = [
+  "https://zfoto.ztek.se/img/full/20230902-DH6_7840.jpg",
+  "https://zfoto.ztek.se/img/full/20230908-DH8_8364.jpg",
+  "https://zfoto.ztek.se/img/full/20230904-DH8_8302.jpg",
+];
