+++ src/data/home-page-select-links.ts
@@ -1,0 +1,16 @@
+export type HomePageSelectLinkType = {
+  title: string;
+  url: string;
+};
+
+export const homePageSelectLinks: HomePageSelectLinkType[] = [
+  {
+    title: "Canvas",
+    url: "https://canla.portal.chalmers.se/canvaslogin/discovery.html?v=1",
+  },
+  {
+    title: "Ladok",
+    url: "https://www.student.ladok.se/student/app/studentwebb/",
+  },
+  { title: "Activity@Z", url: "https://www.facebook.com/groups/activityatz" },
+];
