+++ src/data/home-page-sponsors.ts
@@ -1,0 +1,9 @@
+export type HomePageSponsorType = {
+  company: string;
+  href: string;
+  img: string;
+};
+
+export const homePageSponsors: HomePageSponsorType[] = [
+  { company: "Cpac", img: "/cpac.png", href: "https://cpacsystems.se/" },
+];
